Add experimental support for pasting HTML fragments
This is the first step towards supporting paste with
style information, as well as graphics etc. There is
significant work left to stabilise and improve this
functionality however, so the feature is inactive by
default.

Future work will expand the nodes that can be adopted
and properly synchronise style data.

- Unit tests for HTML adoption code
- Unit tests for HTML text paste