Clean-up hyperlink controller API & add unit tests
Distinguish between inserting a new hyperlink with text, or modifying
the hyperlink for a current selection helps clarify which should be used
at which time.

Added a parameter to the removeHyperlinks function to allow consumers to
specify whether to remove hyperlinks on just the selected text, or
remove
any hyperlink that intersects with the current selection.