Catch any subscriber exceptions when emitting queued signals
Exceptions thrown during the processing of queued signals may improperly
leave the pendingSignals array non-empty. Rather than letting these
exceptions become fatal, simply catch and log, and carry on processing.

The assumption here is that consumers of signals should NOT cause the
processing of an operation to fail (e.g., if a UI binding goes bad).